Add geolocation capability and improve program structure
This commit adds HTML geolocation capability to the program so that each time the web page is refreshed, the user location is centered and shown on Kakao Map. Due to the built-in structure of the geolocation module, the program has been massively edited, mostly due to scope issues. update.js is treated more as the main script and kakao.js is treated more as a module. Documentation in both files are also updated to explain specifically what is happening.