Fix minor issues in ShowMap and showLocationMarker
The mapContainer variable within 'ShowMap' is no longer hard coded within the function and instead defined as a variable in update.js and taken as a argument of showMap. 'showLocationMarker(map, latitude, longitude)'' was also mistakenly documented as a private function while it is also an export function. This commit fixes these minor issues.