Clear markers when pressing button again
This commit solves the issue of markers drawn over when pressing the button again. Many functions in kakao.js have been edited largely to create a clearMarker function, but was eventually discarded because once the button is pressed again, a new marker list is created and this list cannot clear previously drawn markers anyways. Instead, the map is refreshed completely when the button is pressed.