Prevent map overwrite when clicking button
Previous commits aimed to solve the marker overwrite problem by showing a new Kakao Map when the button is clicked. However, this overwrites a new map each time the button is clicked, so this commit solves this by completely clearing the element when the button is clicked and then initialize Kakao Map.