Update Mapo-gu stations (provisional) and introduce async function
Mapo-gu stations are updated in update.js. However, the actual 마포구++ option within the dropdown menu includes a multitude of areas as a test. In the process, it is revealed that the current approach of setTimeout() for drawing markers is too short. However, simply increasing the waiting time is a waste since smaller areas do not need to wait as much. Instead, the button function is changed to an async function where the function waits until update() returns the stations object. Note that the stations object comprises locationsGreen, coordinateValuesGreen, locationsYellow, coordinateValuesYellow and is now a local variable instead of global.