Map re-centers at selected area when button is clicked
This commit adds a few parameters to be used when drawing Kakao Map. Instead of always centering on the user's current location, the map will center on specific coordinates at certain zoom-levels so that the entire area can be visible when the map is redrawn when the button is clicked. In future development, these coordinates and zoom levels can be moved to stationData.js.